feat(IOInputField): add support for fileLoader input type and IOFileInput component
The IOInputField component now supports a new input type called "fileLoader". When the input type is "fileLoader", the component renders an IOFileInput component. The IOFileInput component takes a field prop which is used to set the initial value of the file path. The updateValue function is passed to the IOFileInput component to update the value of the file path in the node's template.