refactor: Update build_inputs method in Component class
The build_inputs method in the Component class has been updated to handle the user_id parameter and return a list of inputs. This change improves the functionality and flexibility of the custom component.

Note: The commit message has been generated based on the provided code changes and recent commits.