refactor: parameterComponent in GenericNode (#3220)
* refactor: create NodeOutputField component in GenericNode

* feat: update isUserChange parameter in setNode function

The `isUserChange` parameter in the `setNode` function of the `FlowStoreType` interface has been updated to be optional. This change improves the flexibility and usability of the function.

* feat: Add NodeInputField component to GenericNode

The NodeInputField component has been added to the GenericNode component. This component is responsible for rendering input fields for nodes in the UI. It improves the functionality and user experience of the GenericNode component.

* delete parameter component

* [autofix.ci] apply automated fixes

* feat: Refactor NodeOutputField component in GenericNode

The NodeOutputField component in the GenericNode component has been refactored to improve code organization and maintainability. This refactor enhances the functionality and user experience of the GenericNode component.

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>