refactor: Minor style fixes (#4578)
* refactor: Update AccountMenu and Select components

- Refactor AccountMenu component to improve readability and simplify logic
- Update Select component to add support for direction prop and display ChevronUp icon when direction is set to "up"

Related work items: #4569, #4565

* refactor: Update header component UI

This commit refactors the header component in the MainPage to update its UI. It adds a sliding indicator for the view type (list or grid) and adjusts the styling of the buttons accordingly. It also removes commented out code related to the store button. Additionally, the commit updates the emptyFolder component by simplifying the text for the "New Flow" button.

Refactor the header component UI and update emptyFolder component

* refactor: Update HeaderMenu component UI

* Updated colors to use shadcn variables

---------

Co-authored-by: Lucas Oliveira <lucas.edu.oli@hotmail.com>