Refactor: Update toTitleCase function and CrashErrorComponent (#4940)
* Refactor: Update Data class to handle different types of text inputs

The Data class in the langflow schema has been updated to handle different types of text inputs. Previously, the page_content attribute was expected to be a string, but now it can also accept other types. If the input is already a string, it is used as is. Otherwise, it is converted to a string before being assigned to the page_content attribute of the Document object.

This change improves the flexibility and robustness of the Data class, allowing it to handle a wider range of input types.

* Refactor: Update toTitleCase function to handle ignoreTitleCase option

* Refactor: Update CrashErrorComponent to add ignoreTitleCase option to the Report on GitHub button

* [autofix.ci] apply automated fixes